Remove dtype argument

+++ packages/tasks/src/model-libraries-snippets.ts
@@ -359,7 +359,7 @@ const diffusersVideoDefaultPrompt = "A man with short gray hair plays a red elec
 const diffusers_default = (model: ModelData) => [
 	`from diffusers import DiffusionPipeline
 
-pipe = DiffusionPipeline.from_pretrained("${model.id}", dtype=torch.bfloat16)
+pipe = DiffusionPipeline.from_pretrained("${model.id}")
 
 prompt = "${get_prompt_from_diffusers_model(model) ?? diffusersDefaultPrompt}"
 image = pipe(prompt).images[0]`,
